Add timetable Array show acts
