+++ .eslintrc
@@ -1,0 +1,12 @@
+{
+  "ecmaFeatures": {
+    "modules": true,
+    "spread": true,
+    "restParams": true
+  },
+  "env": { "es6": true },
+  "globals": {
+    "cy": true
+  },
+  "extends": ["react-app", "plugin:prettier/recommended"]
+}
